requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.4
Django==4.2.3
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
sqlparse==0.4.4

djangorestframework~=3.14.0